pyproject.toml
[project]
name = "akoma-ntoso"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "instructor>=1.11.3",
    "openai>=1.107.2",
    "pdfplumber>=0.10.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
]
